+++ ComicLibrary/Properties/Resources.Designer.cs
@@ -286,6 +286,15 @@ namespace ComicLibrary.Properties {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to Cover variant.
+        /// </summary>
+        public static string CoverVariant {
+            get {
+                return ResourceManager.GetString("CoverVariant", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to Automatically create a backup when saving a file..
         /// </summary>
         public static string CreateBackupWhenSaving {